moved emoji library to diff folder
