+++ packages/nodes/emoji/src/utils/EmojiLibrary/EmojiInlineLibrary.ts
