+++ packages/nodes/emoji/src/utils/EmojiLibrary/EmojiInlineLibrary.types.ts
