+++ packages/nodes/emoji/src/utils/EmojiLibrary/EmojiLibrary.types.ts
