+++ packages/nodes/emoji/src/utils/IndexSearch/index.ts
@@ -5,4 +5,3 @@
 export * from './IndexSearch';
 export * from './EmojiInlineIndexSearch';
 export * from './EmojiFlyoutIndexSearch';
-export * from './EmojiLibrary';
